fix: consolidate projections behavior when deleting records (#1002)
* fix: WR-738 don't fail when nationalroad projection tries to delete something that's already deleted

* fix: consolidate projections behavior when deleting records

+++ src/RoadRegistry.Product.Projections/RoadNodeRecordProjection.cs
@@ -134,7 +134,9 @@ public class RoadNodeRecordProjection : ConnectedProjection<ProductContext>
     private static async Task RemoveRoadNode(ProductContext context, RoadNodeRemoved node, CancellationToken token)
     {
         var roadNode = await context.RoadNodes.FindAsync(node.Id, cancellationToken: token).ConfigureAwait(false);
-
-        context.RoadNodes.Remove(roadNode);
+        if (roadNode is not null)
+        {
+            context.RoadNodes.Remove(roadNode);
+        }
     }
 }
