+++ src/RoadRegistry.Product.Projections/RoadSegmentNumberedRoadAttributeRecordProjection.cs
@@ -10,6 +10,7 @@ using BackOffice.Extracts.Dbase.RoadSegments;
 using BackOffice.Messages;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.Connector;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.SqlStreamStore;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.IO;
 using Schema;
 
@@ -68,7 +69,7 @@ public class RoadSegmentNumberedRoadAttributeRecordProjection : ConnectedProject
                         await RoadSegmentRemovedFromNumberedRoad(context, numberedRoad, token);
                         break;
                     case RoadSegmentRemoved roadSegmentRemoved:
-                        RoadSegmentRemoved(context, roadSegmentRemoved);
+                        await RoadSegmentRemoved(context, roadSegmentRemoved, token);
                         break;
                 }
         });
@@ -129,14 +130,14 @@ public class RoadSegmentNumberedRoadAttributeRecordProjection : ConnectedProject
         }.ToBytes(manager, encoding);
     }
 
-    private static void RoadSegmentRemoved(ProductContext context, RoadSegmentRemoved roadSegmentRemoved)
+    private static async Task RoadSegmentRemoved(ProductContext context, RoadSegmentRemoved roadSegmentRemoved, CancellationToken token)
     {
         var segmentNumberedRoadAttributeRecords =
-            context.RoadSegmentNumberedRoadAttributes
-                .Local
+            context.RoadSegmentNumberedRoadAttributes.Local
                 .Where(x => x.RoadSegmentId == roadSegmentRemoved.Id)
-                .Concat(context.RoadSegmentNumberedRoadAttributes
-                    .Where(x => x.RoadSegmentId == roadSegmentRemoved.Id));
+                .Concat(await context.RoadSegmentNumberedRoadAttributes
+                    .Where(x => x.RoadSegmentId == roadSegmentRemoved.Id)
+                    .ToArrayAsync(token));
 
         context.RoadSegmentNumberedRoadAttributes.RemoveRange(segmentNumberedRoadAttributeRecords);
     }
@@ -147,7 +148,9 @@ public class RoadSegmentNumberedRoadAttributeRecordProjection : ConnectedProject
     {
         var roadSegment =
             await context.RoadSegmentNumberedRoadAttributes.FindAsync(numberedRoad.AttributeId, cancellationToken: token).ConfigureAwait(false);
-
-        context.RoadSegmentNumberedRoadAttributes.Remove(roadSegment);
+        if (roadSegment is not null)
+        {
+            context.RoadSegmentNumberedRoadAttributes.Remove(roadSegment);
+        }
     }
 }
